+++ R/parse.R
@@ -144,7 +144,7 @@ parse_idd_file <- function(path, epw = FALSE) {
 
     # complete property columns
     dt_field <- complete_property(dt_field, "field", dt_class, epw = epw)
-    dt_class <- complete_property(dt_class, "class", dt_field)
+    dt_class <- complete_property(dt_class, "class", dt_field, epw = epw)
 
     # ConnectorList references are missing until v9.1
     # See https://github.com/NREL/EnergyPlus/issues/7172
@@ -746,7 +746,9 @@ complete_property <- function (dt, type, ref, epw = FALSE) {
     }
 
     # add missing property columns if necessary
-    for (key in unlist(keys, use.names = FALSE)) {
+    cols <- unlist(keys, use.names = FALSE)
+    if (!epw) cols <- setdiff(cols, c("missing", "exist-minimum", "exist-minimum>", "exist-maximum", "exist-maximum<"))
+    for (key in cols) {
         if (!has_names(dt, key)) set(dt, NULL, key, slash_init_value(key))
     }
 
@@ -1095,6 +1097,7 @@ sep_idf_lines <- function (dt, type_enum) {
         body = stri_trim_right(stri_sub(string, to = excl_loc - 1L)),
         comment = stri_sub(string, excl_loc + 1L)
     )]
+    dt[excl_loc <= spcl_loc, `:=`(excl_loc = NA_integer_)]
     dt[is.na(excl_loc) & !is.na(spcl_loc), `:=`(comment = stri_trim_left(stri_sub(comment, 2L)))]
 
     set(dt, NULL, c("excl_loc", "spcl_loc"), NULL)
@@ -1201,7 +1204,7 @@ sep_header_options <- function (dt, type_enum) {
 
     if (!length(out)) return(res)
 
-    save_format <- c("sorted", "ori_top", "ori_bot")
+    save_format <- c("sorted", "new_top", "new_bot")
     for (lgl in setdiff(out, save_format)) res$options[[lgl]] <- TRUE
     sf <- save_format[save_format %in% out][1L]
     if (!is.na(sf)) res$options$save_format <- sf